Moves ad tests one level out of src
