+++ packages/ad/__tests__/ad-init-grapeshot.test.js
@@ -1,4 +1,4 @@
-import adInit from "../ad-init";
+import adInit from "../src/ad-init";
 import { makeAdInitMocks } from "./ad-init-mocks";
 
 describe("AdInit.grapeshot", () => {
