+++ packages/ad/__tests__/ad.test.js
@@ -3,15 +3,15 @@ import renderer from "react-test-renderer";
 import Enzyme from "enzyme";
 import React16Adapter from "enzyme-adapter-react-16";
 
-import Ad, { AdComposer } from "../ad";
+import Ad, { AdComposer } from "../src/ad";
 
 Enzyme.configure({ adapter: new React16Adapter() });
 
 // prevent function sources appearing in snapshots
-jest.mock("../webview-event-callback-setup", () => "mockErrorHandler");
-jest.mock("../ad-init", () => () => "mockInit");
+jest.mock("../src/webview-event-callback-setup", () => "mockErrorHandler");
+jest.mock("../src/ad-init", () => () => "mockInit");
 
-jest.mock("../placeholder", () => "Placeholder"); // prevent SVG in snapshots
+jest.mock("../src/placeholder", () => "Placeholder"); // prevent SVG in snapshots
 jest.mock("WebView", () => "WebView"); // https://github.com/facebook/react-native/issues/12440
 
 describe("Ad", () => {
