+++ packages/ad/__tests__/check-self-contained-function.js
