+++ packages/ad/__tests__/dom-context-stories.test.js
@@ -1,4 +1,4 @@
-import domContextInit from "../../ad.stories-domcontext-init";
+import domContextInit from "../ad.stories-domcontext-init";
 import { expectFunctionToBeSelfContained } from "./check-self-contained-function";
 
 describe("DOMContext Stories", () => {
