+++ packages/ad/__tests__/dom-context.native.test.js
@@ -2,10 +2,10 @@ import React from "react";
 import { Linking } from "react-native";
 import renderer from "react-test-renderer";
 
-import DOMContextNative from "../dom-context";
+import DOMContextNative from "../src/dom-context";
 
 // prevent function sources appearing in snapshots
-jest.mock("../webview-event-callback-setup", () => "mockErrorHandler");
+jest.mock("../src/webview-event-callback-setup", () => "mockErrorHandler");
 
 jest.mock("WebView", () => "WebView"); // https://github.com/facebook/react-native/issues/12440
