+++ packages/ad/__tests__/placeholder.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 
-import Placeholder from "../placeholder";
+import Placeholder from "../src/placeholder";
 
 jest.mock("@times-components/watermark", () => () => "watermark");
