+++ packages/ad/package.json
@@ -22,7 +22,7 @@
       "node_modules/(?!(react-native|@times-components)/)"
     ],
     "testMatch": [
-      "<rootDir>/packages/ad/src/__tests__/**.test.js"
+      "<rootDir>/packages/ad/__tests__/**.test.js"
     ],
     "testEnvironment": "jsdom"
   },
